backend/parser/requirements.txt
weaviate-client==4.14.1
requests
yake
nltk
pymorphy3
razdel